Remove the child nodes from CloudFromation Stack Node in th explorer
* We only supported Lambda and adding new ones slows down the list since each one needs an additional network call per node, so remove it since the benefit of having it was small
* Remove the idea of "immutable" versions of explorer nodes
* This also makes it a leaf node so we can add double click to open stack in the future
