+++ .changes/next-release/breaking-59cde94b-5517-4f68-a5c1-ea5be2702321.json
@@ -1,0 +1,4 @@
+{
+  "type" : "breaking",
+  "description" : "Remove the Lambda nodes underneath of the CloudFromation stack in the explorer"
+}
