+++ jetbrains-core/resources/META-INF/plugin.xml
@@ -252,15 +252,11 @@
             <action id="lambda.function.create" class="software.aws.toolkits.jetbrains.services.lambda.upload.CreateLambdaFunction"/>
         </group>
 
-        <group id="aws.toolkit.explorer.lambda.function.immutable" popup="false" compact="false">
+        <group id="aws.toolkit.explorer.lambda.function" popup="false" compact="false">
             <reference ref="RunContextGroupInner"/>
             <separator/>
             <action id="lambda.function.gotohandler"
                     class="software.aws.toolkits.jetbrains.services.lambda.actions.GoToHandlerAction"/>
-        </group>
-
-        <group id="aws.toolkit.explorer.lambda.function" popup="false" compact="false">
-            <reference ref="aws.toolkit.explorer.lambda.function.immutable"/>
             <separator/>
             <action id="lambda.function.updateConfiguration"
                     class="software.aws.toolkits.jetbrains.services.lambda.actions.UpdateFunctionConfigurationAction"/>
@@ -275,7 +271,7 @@
                     class="software.aws.toolkits.jetbrains.core.stack.OpenStackUiAction"/>
             <separator/>
             <action id="cloudformation.stack.delete"
-                    class="software.aws.toolkits.jetbrains.services.cloudformation.DeleteCloudFormationStackAction"/>
+                    class="software.aws.toolkits.jetbrains.services.cloudformation.actions.DeleteStackAction"/>
         </group>
 
         <group id="aws.toolkit.explorer.schemas" popup="true" compact="false">
