+++ jetbrains-core/src/software/aws/toolkits/jetbrains/core/explorer/nodes/AwsExplorerResourceNode.kt
@@ -8,24 +8,15 @@ import javax.swing.Icon
 
 /**
  * Top level class for a node that represents a resource such as an AWS Lambda.
- *
- * @param immutable Used to indicate that the resource should not support CRUD actions since it is managed by something like CloudFormation
  */
 abstract class AwsExplorerResourceNode<T>(
     project: Project,
     private val serviceId: String,
     value: T,
-    awsIcon: Icon,
-    private val immutable: Boolean = false
+    awsIcon: Icon
 ) : AwsExplorerNode<T>(project, value, awsIcon),
     ResourceActionNode {
-    override fun actionGroupName() = "aws.toolkit.explorer.$serviceId.${resourceType()}".let {
-        if (immutable) {
-            "$it.immutable"
-        } else {
-            it
-        }
-    }
+    override fun actionGroupName() = "aws.toolkit.explorer.$serviceId.${resourceType()}"
 
     override fun isAlwaysLeaf(): Boolean = true
