+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/cloudformation/resources/CloudFormationResources.kt
@@ -4,7 +4,6 @@
 package software.aws.toolkits.jetbrains.services.cloudformation.resources
 
 import software.amazon.awssdk.services.cloudformation.CloudFormationClient
-import software.amazon.awssdk.services.cloudformation.model.StackResourceSummary
 import software.amazon.awssdk.services.cloudformation.model.StackStatus
 import software.amazon.awssdk.services.cloudformation.model.StackSummary
 import software.aws.toolkits.jetbrains.core.ClientBackedCachedResource
@@ -19,9 +18,4 @@ object CloudFormationResources {
 
     @JvmField
     val ACTIVE_STACKS = LIST_STACKS.filter { it.stackStatus() != StackStatus.DELETE_COMPLETE }.filter { it.stackName() != null }
-
-    fun listStackResources(stackId: String): Resource.Cached<List<StackResourceSummary>> =
-        ClientBackedCachedResource(CloudFormationClient::class, "cloudformation.list_resources.$stackId") {
-            listStackResourcesPaginator { it.stackName(stackId) }.stackResourceSummaries().toList()
-        }
 }
